backend/requirements.txt
fastapi
uvicorn
anthropic
python-dotenv
langchain
langgraph
langchain-anthropic==0.3.13